requirements.txt
pytest==7.4.3
request==2.28.1